move all data logic to :common module
