+++ common/src/commonMain/kotlin/work/racka/template/common/sources/DbClient.kt
@@ -1,0 +1,5 @@
+package work.racka.template.common.sources
+
+internal expect class DbClient {
+    fun platformName(): String
+}
