+++ common/src/commonMain/kotlin/work/racka/template/common/sources/MyRepository.kt
@@ -1,14 +1,13 @@
-package dependencies
+package work.racka.template.common.sources
 
 interface MyRepository {
     fun helloWorld(): String
 }
 
-class MyRepositoryImpl(
+internal class MyRepositoryImpl(
     private val dbClient: DbClient
 ): MyRepository {
     override fun helloWorld(): String {
-        return "Hello World"
+        return "Hello: ${dbClient.platformName()}"
     }
-
 }
