+++ common/src/desktopMain/kotlin/work/racka/template/common/di/Common.desktop.kt
@@ -1,7 +1,10 @@
 package work.racka.template.common.di
 
 import org.koin.core.module.Module
+import org.koin.core.module.dsl.singleOf
+import org.koin.dsl.module
+import work.racka.template.common.sources.DbClient
 
-internal actual fun platform(): Module {
-    TODO("Not yet implemented")
+internal actual fun platform(): Module = module {
+    singleOf(::DbClient)
 }
