+++ common/src/desktopMain/kotlin/work/racka/template/common/sources/DbClient.desktop.kt
@@ -1,0 +1,5 @@
+package work.racka.template.common.sources
+
+internal actual class DbClient {
+    actual fun platformName(): String = "Java ${System.getProperty("java.version")}"
+}
