--- composeApp/src/androidMain/kotlin/Platform.android.kt
@@ -1,7 +1,0 @@
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
