--- composeApp/src/androidMain/kotlin/dependencies/DbClient.android.kt
@@ -1,7 +1,0 @@
-package dependencies
-
-import android.content.Context
-
-actual class DbClient(
-    private val context:Context
-)
