--- composeApp/src/commonMain/kotlin/Greeting.kt
@@ -1,7 +1,0 @@
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
