--- composeApp/src/commonMain/kotlin/dependencies/DbClient.kt
@@ -1,3 +1,0 @@
-package dependencies
-
-expect class DbClient
