--- composeApp/src/desktopMain/kotlin/dependencies/DbClient.desktop.kt
@@ -1,3 +1,0 @@
-package dependencies
-
-actual class DbClient
