--- composeApp/src/desktopMain/kotlin/di/Modules.desktop.kt
@@ -1,10 +1,0 @@
-package di
-
-import dependencies.DbClient
-import org.koin.core.module.Module
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.module
-
-internal actual val platformModule: Module = module {
-    singleOf(::DbClient)
-}
